+++ app/src/main/kotlin/com/architects/pokearch/core/data/datasource/PokemonLocalDataSource.kt
@@ -16,7 +16,7 @@ interface PokemonLocalDataSource {
 
     suspend fun getPokemonInfo(id: Int): PokemonInfo?
 
-    suspend fun numCount(): Int
+    suspend fun countPokemon(): Int
 
     suspend fun randomId(): Int
 }
