+++ app/src/main/kotlin/com/architects/pokearch/core/framework/database/entities/converters/TypesHolderConverter.kt
@@ -1,4 +1,4 @@
-package com.architects.pokearch.core.framework.local.entities.converters
+package com.architects.pokearch.core.framework.database.entities.converters
 
 import androidx.room.TypeConverter
 import com.google.gson.Gson
