+++ app/src/main/kotlin/com/architects/pokearch/core/framework/network/di/annotations/PokeRetrofit.kt
@@ -1,0 +1,7 @@
+package com.architects.pokearch.core.framework.network.di.annotations
+
+import javax.inject.Qualifier
+
+@Retention(AnnotationRetention.RUNTIME)
+@Qualifier
+annotation class PokeRetrofit
