+++ app/src/main/kotlin/com/architects/pokearch/core/framework/sensors/AndroidSensor.kt
@@ -5,6 +5,7 @@ import android.hardware.Sensor
 import android.hardware.SensorEvent
 import android.hardware.SensorEventListener
 import android.hardware.SensorManager
+import com.architects.pokearch.core.data.datasource.SensorSource
 
 abstract class AndroidSensor(
     private val context: Context,
