+++ app/src/main/kotlin/com/architects/pokearch/ui/features/home/ui/HomeItem.kt
@@ -34,8 +34,8 @@ import com.architects.pokearch.ui.components.image.ArchAsyncImage
 
 @Composable
 fun HomeItem(
-    modifier: Modifier = Modifier,
     pokemon: Pokemon,
+    modifier: Modifier = Modifier,
     onItemClick: (Int) -> Unit,
 ) {
